Another bug uncovered by Chaewon's Crit tests in swlogiclab/checker-framework-effects#31
The framework hadn't really been tested with an EQ that had partial sequencing yet, so in the method invocation case, if any effect push was undefined that wasn't caught, and instead a subsequent squash would fail because errors from the failing seq call weren't pulled after the failure.  Now we check for failure, and issue a warning if necessary. There are likely other examples of this with other AST nodes, but we won't find them until we run on more code.  In the mean time, this test can't even be tested without adding a new EQ with partial seq, but apparently the incoming Crit tests contain such a case.